backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
supabase==2.10.0
transformers==4.47.1
sentence-transformers==3.3.1
torch>=2.0.0
torchaudio>=2.0.0
numpy==1.26.4
httpx==0.27.2
pydantic==2.10.3
python-multipart==0.0.20
